Refactor core unit tests: break monolithic files into focused modules
- Split test_extractor.py into 4 focused modules:
  * test_extractor_path_resolution.py - Path resolution logic
  * test_extractor_checksum.py - Checksum calculations
  * test_extractor_extraction.py - File/directory extraction
  * test_extractor_validation.py - Library validation & management

- Split test_installer.py into 4 focused modules:
  * test_installer_config.py - Configuration & lockfile operations
  * test_installer_single.py - Single library installation
  * test_installer_batch.py - Batch installation operations
  * test_installer_management.py - Library management operations

Benefits: Easier debugging, focused testing, cleaner git diffs
Status: Files split but need breaking change fixes from metadata refactor
